First round of getting rid of warnings caused by implicit conversions

+++ Source/Utilities/BufferQueue.h
@@ -31,9 +31,9 @@ public:
         jassert (samplesPerBuffer > 0);
         jassert (numBuffers > 0);
 
-        buffers.resize (numBuffers);
-        for (int i = 0; i < numBuffers; ++i)
-            buffers[i].resize (samplesPerBuffer);
+        buffers.resize (uint(numBuffers));
+        for (uint i = 0; i < uint(numBuffers); ++i)
+            buffers[i].resize (uint(samplesPerBuffer));
 
         fifo.setTotalSize (numBuffers);
     }
@@ -70,10 +70,10 @@ public:
         auto scopedWrite = fifo.write (1);
 
         jassert (numSamples <= bufferSize);
-        juce::FloatVectorOperations::copy (buffers[scopedWrite.startIndex1].data(), dataToPush, juce::jmin (bufferSize, numSamples));
+        juce::FloatVectorOperations::copy (buffers[uint(scopedWrite.startIndex1)].data(), dataToPush, juce::jmin (bufferSize, numSamples));
 
         if (numSamples < bufferSize)
-            juce::FloatVectorOperations::clear (buffers[scopedWrite.startIndex1].data() + numSamples, bufferSize - numSamples);
+            juce::FloatVectorOperations::clear (buffers[uint(scopedWrite.startIndex1)].data() + numSamples, bufferSize - numSamples);
 
         return true;
     }
@@ -86,7 +86,7 @@ public:
 
         auto scopedRead = fifo.read (1);
 
-        juce::FloatVectorOperations::copy (outputBuffer, buffers[scopedRead.startIndex1].data(), bufferSize);
+        juce::FloatVectorOperations::copy (outputBuffer, buffers[uint(scopedRead.startIndex1)].data(), bufferSize);
 
         return true;
     }
